[최적화] NowPlaying BPM 시간계산 방식 수정 - 매 프레임 TIming 리스트 순회하며 계산했던 것에서 지나간 타이밍 저장해두고 현재 타이밍에 대해서만 계산하도록 개선
[UI] InGame 패널 및 Text크기 조절, 노트 크기 증가

[구조개선] CustomLayoutGroup spacing, padding int -> float 변경

+++ Assets/Scripts/UISys/LayoutGroup/CustomLayoutGroup.cs
@@ -9,17 +9,17 @@ using UnityEngine.UI;
 [RequireComponent( typeof( RectTransform ) )]
 public abstract class CustomLayoutGroup : MonoBehaviour, ILayoutController
 {
-    public int spacing;
+    public float spacing;
     [Serializable]
     protected struct RectOffset
     {
-        public int left, right, top, bottom;
+        public float left, right, top, bottom;
 
-        public RectOffset( int _left, int _right, int _top, int _bottom )
+        public RectOffset( float _left, float _right, float _top, float _bottom )
         {
-            left = _left;
-            right = _right;
-            top = _top;
+            left   = _left;
+            right  = _right;
+            top    = _top;
             bottom = _bottom;
         }
     }
